p9221_charger: Increase BPP current limit to 900 mA
The BPP (Baseline Power Profile) specification states that the maximum
power is 5 W. The p9221 charger supports 900 mA for BPP chargers, but the
BPP current limit is set to 700 mA for some reason. This is especially odd
given that the BPP ramp-up current limit (P9221_DC_ICL_BPP_RAMP_DEFAULT_UA)
is set correctly to 900 mA.

Fix the slow BPP charging rate by raising the BPP current limit to 900 mA.
This is evidenced to be the maximum BPP rate given OVC_BACKOFF_LIMIT and
the presence of over-current interrupts when attempting to use 1000 mA.

For the current increase to work consistently, the ICL ramp-up delay must
be increased, otherwise wireless charging may cut out when ramping up on a
BPP charger. Increase the ramp-up delay from ~100 ms to 500 ms for this
reason, which has the added benefit of making wireless charging overall
more reliable.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>

Change-Id: Ie42ff0a255f87dc0aeb1ebe2858599a31122f8c4